Set the warning level to 4 for the ResText, TortoiseMerge, TortoiseBlame and SubWCRev project and get rid of all the warnings.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@1905 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/ResText/ResModule.h
@@ -90,5 +90,5 @@ private:
 	int				m_bTranslatedMenuStrings;
 	int				m_bDefaultMenuStrings;
 	
-	int				m_wTargetLang;
+	WORD			m_wTargetLang;
 };
